Extract functionality to `lang::ir::utils`
